feat(crypt): fix bug in Copy and Move action
